- Added questions and answers

+++ src/main/java/com/litequizapp/service/CategoryRestService.java
@@ -1,5 +1,8 @@
-package com.litequizapp.server;
+package com.litequizapp.service;
 
+import com.litequizapp.entity.CategoryEntity;
+import com.litequizapp.exception.ElementNotFoundException;
+import com.litequizapp.repository.CategoryRepository;
 import java.util.ArrayList;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,11 +22,14 @@ public class CategoryRestService {
 
   public String getCategoryById(long id) {
     CategoryEntity category = categoryRepository.findById(id);
+    if (category == null) {
+      throw new ElementNotFoundException();
+    }
     return category.toString();
 
   }
 
-  public List<String> getAllCategory() {
+  public List<String> getAllCategories() {
     List<String> categories = new ArrayList<>();
     for (CategoryEntity category : categoryRepository.findAll()) {
       categories.add(category.toString());
@@ -38,11 +44,18 @@ public class CategoryRestService {
 
   public void updateCategory(long id, String title) {
     CategoryEntity category = categoryRepository.findById(id);
+    if (category == null) {
+      throw new ElementNotFoundException();
+    }
     category.setTitle(title);
     categoryRepository.save(category);
   }
 
   public void deleteCategory(long id) {
+    CategoryEntity category = categoryRepository.findById(id);
+    if (category == null) {
+      throw new ElementNotFoundException();
+    }
     categoryRepository.deleteById(id);
 
   }
